removed various requires of spec_helper
Because the rails:rspec generator adds a require of the spec_helper to
the project's .rspec file, there is never any need to require it in any
spec files.

Also, since the spec directory has been added to the Ruby load path, we
don't need to path to the rails_helper when requiring it.